requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
astroid==3.0.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.1
dataclasses-json-speakeasy==0.5.11
Deprecated==1.2.14
dill==0.3.7
emoji==2.10.1
et-xmlfile==1.1.0
faiss-cpu==1.7.4
fastapi==0.109.0
filetype==1.2.0
Flask==3.0.0
Flask-Limiter==3.5.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.40
greenlet==3.0.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
langchain==0.0.319
langchain-community==0.0.17
langchain-core==0.1.18
langdetect==1.0.9
langsmith==0.0.86
limits==3.7.0
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
openai==0.28.1
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
platformdirs==4.0.0
protobuf==4.24.4
pyarrow==13.0.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pylint==3.0.2
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
pypdf==4.0.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2024.2.7
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.6.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.22
starlette==0.35.1
streamlit==1.27.2
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.1
toml==0.10.2
tomlkit==0.12.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
unstructured==0.12.3
unstructured-client==0.17.0
urllib3==2.0.7
uvicorn==0.27.0
validators==0.22.0
watchdog==3.0.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.2
zipp==3.17.0